feat(user-page): create separate component for username field
